Implements permissions for users with books

+++ ucc/biblioteca/views.py
@@ -2,12 +2,11 @@ from django.shortcuts import redirect, render
 from django.http import HttpResponse
 from django.shortcuts import render
 from django.contrib import messages
+from django.core.exceptions import PermissionDenied
 from .models import Book, Booking, User
+from .utils import validate_session
 
 # Create your views here.
-def signin(request):
-    pass
-
 def login(request):
     if request.method == 'POST':
         email = request.POST.get("email")
@@ -32,14 +31,24 @@ def login(request):
             messages.error(request, "Something way wrong...")
             return redirect("login")
     elif request.method == 'GET':
+        is_logout = validate_session(request)
+        if is_logout is None:
+            return redirect("index")
         return render(request, "auth/login.html", context={'title': 'Login'})
 
 def logout(request):
-    pass
+    try:
+        del request.session["user_logged"]
+        messages.success(request, "Come back soon!")
+        return redirect("login")
+    except:
+        messages.error(request, "Error. Try again...")
+        return redirect("index")
 
 def index(request):
+    is_logout = validate_session(request, "Please log in to access")
     context = {"title": "Home"}
-    return render(request, "index.html", context)
+    return render(request, "index.html", context) if is_logout is None else is_logout
 
 def sell(request):
     context = {"title": "Sell"}
@@ -66,15 +75,20 @@ def sum(request, n1, n2):
 
 # CRUD books
 def book_list(request):
+    is_logout = validate_session(request, "Please log in to access")
     context = {'title': 'Books'}
     context["books"] = Book.objects.all()
-    return render(request, "books/index.html", context)
+    return render(request, "books/index.html", context) if is_logout is None else is_logout
 
 def create_book(request):
+    is_logout = validate_session(request, "Unauthorized. Log in to get access")
+    if is_logout is not None: return is_logout
     if(request.method == 'GET'):
         return render(request, "books/form.html", {'title': "Create Book"})
     elif(request.method == 'POST'):
         try:
+            user_rol = request.session.get("user_logged")["rol"]
+            if user_rol not in ["A", "L"]: raise PermissionDenied()
             book = Book()
             book.title = request.POST.get("title")
             book.author = request.POST.get("author")
@@ -82,6 +96,8 @@ def create_book(request):
             book.publication_date = request.POST.get("publication_date")
             book.save()
             messages.success(request, "Book created!")
+        except PermissionDenied:
+            messages.warning(request, "Unathorize. Contact with Library or Admin for more info")
         except Exception as e:
             print(e)
             messages.error(request, "Something way wrong...")
@@ -89,8 +105,12 @@ def create_book(request):
         return redirect("list_books")
 
 def edit_book(request, id):
+    is_logout = validate_session(request, "Unauthorized. Log in to get access")
+    if is_logout is not None: return is_logout
     if request.method == "POST":
         try:
+            user_rol = request.session.get("user_logged")["rol"]
+            if user_rol not in ["A", "L"]: raise PermissionDenied()
             book = Book.objects.get(pk = id)
             book.title = request.POST.get("title")
             book.author = request.POST.get("author")
@@ -98,6 +118,8 @@ def edit_book(request, id):
             book.publication_date = request.POST.get("publication_date")
             book.save()
             messages.success(request, "Book created!")
+        except PermissionDenied:
+            messages.warning(request, "Unathorize. Contact with Library or Admin for more info")
         except Exception as e:
             print(e)
             messages.error(request, "Something way wrong...")
@@ -108,10 +130,16 @@ def edit_book(request, id):
         return render(request, "books/form.html", context)
 
 def delete_book(request, id):
+    is_logout = validate_session(request, "Unauthorized. Log in to get access")
+    if is_logout is not None: return is_logout
     try:
+        user_rol = request.session.get("user_logged")["rol"]
+        if user_rol not in ["A", "L"]: raise PermissionDenied()
         book = Book.objects.get(pk = id)
         book.delete()
         messages.success(request, "Book deleted!")
+    except PermissionDenied:
+        messages.warning(request, "Unathorize. Contact with Library or Admin for more info")
     except Exception as e:
         print(e)
         messages.error(request, "Something way wrong...")
